+++ src/routes/signin/+page.server.ts
@@ -2,16 +2,16 @@ import { superValidate, message } from 'sveltekit-superforms/server';
 import { zod } from 'sveltekit-superforms/adapters';
 
 import { fail } from '@sveltejs/kit';
-import { schemaLastStep } from './schema.js';
+import { signInFormSchema } from './schema.js';
 
 export const load = async () => {
-  const form = await superValidate(zod(schemaLastStep));
+  const form = await superValidate(zod(signInFormSchema));
   return { form };
 };
 
 export const actions = {
   default: async ({ request }) => {
-    const form = await superValidate(request, zod(schemaLastStep));
+    const form = await superValidate(request, zod(signInFormSchema));
 
     console.log('form actions default:', form);
