+++ src/routes/signin/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { type PageData } from './$types.js';
-  import SignInForm from './sign-in-form2.svelte';
+  import SignInForm from './sign-in-form.svelte';
 
   let { data }: { data: PageData } = $props();
 </script>
