+++ src/routes/signin/schema.ts
@@ -2,62 +2,80 @@ import { z } from 'zod';
 
 export const emailSchema = z.string().email({
   message: 'Please enter a valid email address.',
-}).optional();
+});
 
 export const usernameSchema = z.string({
-  message: 'A username must be at least 3 characters.',
-}).min(3).max(30).optional();
+  message: 'A username must be at least 3 characters.'
+}).min(3).max(30);
 
-export const passwordSchema = z.string().min(8, {
-  message: 'Your password must be at least 8 characters.',
-}).optional();
-
-export const otpSchema = z.string().min(6, {
+const otpSchema = z.string().min(6, {
   message: 'Your one-time password must be at least 6 characters.',
-}).optional();
+}).transform((val) => val.trim());
 
-export const getOtpMessage = (formData: {
-  email?: string;
-  username?: string 
+const passwordSchema = z.string().min(8, {
+  message: 'Your password must be at least 8 characters.',
+}).transform((val) => val.trim());
+
+export const schemaFirstStep = z.object({
+  ident: z
+    .string()
+    .min(3, 'Username or email is required')
+    .transform((val) => val.trim()),
+  // token: z.undefined(),
+  token: otpSchema.optional(),
+  password: passwordSchema.optional(),
+  authType: z.literal('password').default('password'),
+});
+
+export const schemaLastStep = schemaFirstStep.extend({
+  token: otpSchema.optional(),
+  authType: z.literal('token').optional(),
+});
+
+export const signInFormSchema = z.discriminatedUnion('authType', [
+  schemaFirstStep,
+  schemaLastStep,
+]);
+
+export const getOtpMessage = (formData: { ident?: string }) => {
+  const identifier = formData.ident || '';
+  return `Enter the verification code sent to ${identifier}`;
+};
+
+export const shouldUseTokenAuth = (formData: {
+  token?: string;
+  password?: string;
+}): boolean => {
+  return !!formData.token && (!formData.password || formData.password.length === 0);
+};
+
+export const validateAuthForm = (formData: {
+  ident: string;
+  password?: string;
+  token?: string;
+  authType: 'password' | 'token';
 }) => {
-  if (formData.email) {
-    return `Enter the six digit code that was sent to ${formData.email}`;
-  } else if (formData.username) {
-    return `Enter the six digit code that was sent to the email associated with ${formData.username}`;
-  } else {
-    return `Enter the six digit verification code`;
+  try {
+    return {
+      success: true,
+      data: signInFormSchema.parse(formData),
+      error: null,
+    };
+  } catch (error) {
+    if (error instanceof z.ZodError) {
+      return {
+        success: false,
+        data: null,
+        error: error.format(),
+      };
+    }
+    return {
+      success: false,
+      data: null,
+      error: 'An unexpected error occurred',
+    };
   }
 };
 
-export const schemaFirstStep = z.object({
-  email: emailSchema,
-  username: usernameSchema,
-}).refine(
-  data => (data.email && data.email.length > 0) || (data.username && data.username.length > 0),
-  {
-    message: "Either email or username must be provided",
-    path: ["identifierMissing"]
-  }
-);
-
-export const schemaLastStep = z.object({
-  email: emailSchema,
-  username: usernameSchema,
-  password: passwordSchema,
-  emailOtp: otpSchema,
-}).refine(
-  data => data.email || data.username,
-  {
-    message: "Either email or username must be provided",
-    path: ["identifierMissing"]
-  }
-).refine(
-  data => data.password || data.emailOtp,
-  {
-    message: "Either password or one-time password must be provided",
-    path: ["authMethodMissing"]
-  }
-);
-
-// The exported type should be the totality of the form. Since this is stepped, we need to specify the last.
-export type SignInFormSchema = z.infer<typeof schemaLastStep>;
+// Since steps swap required values, we need to join the two possible schemas to account for all possible requirements
+export type SignInFormSchema = z.infer<typeof signInFormSchema>;
